+++ src/components/common/Search.vue
@@ -1,6 +1,6 @@
 <template>
   <div class="search">
-    <div class="relative flex cursor-text" :style="isFocused ? '' : 'background-image: none'" @click="setFocus">
+    <div class="suffix relative flex cursor-text" :style="isFocused ? '' : 'background-image: none'" @click="setFocus">
       <input
         ref="searchInput"
         :value="keyword"
@@ -11,7 +11,6 @@
           w-full
           py-2
           pr-4
-          pl-10
           text-0
           font-normal
           text-text
@@ -24,16 +23,21 @@
           border-none
           appearance-none
         "
-        type="text"
+        :class="[isSearchIconVisible ? 'pl-10' : 'pl-4']"
+        :type="inputType"
         :placeholder="placeholder"
         @input="$emit('update:keyword', $event.target.value)"
+        @focus="$emit('focus:value', $event.target.value)"
+        @blur="$emit('blur:value', $event.target.value)"
       />
       <Icon
+        v-if="isSearchIconVisible"
         class="icon-search absolute z-10 h-full px-3 text-muted pointer-events-none"
         :name="'MagnifyingGlassIcon'"
         :icon-size="1"
       />
       <Icon
+        v-if="isCloseIconVisible"
         v-show="keyword !== ''"
         class="
           icon-reset
@@ -56,7 +60,11 @@
           }
         "
       />
-      <div class="focus-border absolute z-0 -inset-0.5 rounded-xl invisible bg-gold-circular" />
+      <slot></slot>
+      <div
+        class="focus-border absolute z-0 -inset-0.5 rounded-xl invisible"
+        :class="[borderColour ? borderColour : 'bg-gold-circular']"
+      />
     </div>
   </div>
 </template>
@@ -68,11 +76,15 @@ export default defineComponent({
   name: 'Search',
   components: { Icon },
   props: {
-    keyword: { type: String, required: false, default: null },
+    keyword: { type: [String, Number], required: false, default: null },
     placeholder: { type: String, required: false, default: 'Search' },
     autofocus: { type: Boolean, default: true },
+    isSearchIconVisible: { type: Boolean, default: true },
+    isCloseIconVisible: { type: Boolean, default: true },
+    inputType: { type: String, default: 'text' },
+    borderColour: { type: String, default: null },
   },
-  emits: ['update:keyword'],
+  emits: ['update:keyword', 'blur:value', 'focus:value'],
   setup(props) {
     const searchInput = ref(null);
     const isFocused = ref(false);