+++ src/components/stake/SwitchForm/index.ts
@@ -1,0 +1,3 @@
+import SwitchForm from './SwitchForm.vue';
+
+export default SwitchForm;