+++ src/views/Asset.vue
@@ -69,7 +69,13 @@
               </dd>
             </div>
 
-            <div>
+            <div v-if="assetConfig?.stakable && stakingEnabled">
+              <dt class="text-muted">{{ $t('pages.asset.unbonding') }}</dt>
+              <dd class="font-medium mt-0.5">
+                <AmountDisplay :amount="{ amount: unstakedAmount, denom }" />
+              </dd>
+            </div>
+            <div v-else>
               <dt class="text-muted">{{ $t('pages.asset.pooled') }}</dt>
               <dd class="font-medium mt-0.5">
                 <tippy>
@@ -125,6 +131,13 @@
           </ul>
         </section>
 
+        <!-- Staking -->
+        <template v-if="stakingEnabled">
+          <section v-if="assetConfig?.stakable" class="mt-16">
+            <StakeTable class="mt-8" :denom="denom" />
+          </section>
+        </template>
+
         <!-- Pools -->
 
         <section v-if="poolsDisplay.length" class="mt-16">
@@ -139,14 +152,6 @@
           </header>
 
           <Pools :pools="poolsDisplay" class="mt-8" />
-        </section>
-
-        <!-- Staking -->
-
-        <section v-if="assetConfig?.stakable" class="mt-16">
-          <h2 class="text-2 font-bold">{{ $t('pages.asset.staking') }}</h2>
-
-          <StakeTable class="mt-8" :denom="denom" />
         </section>
       </main>
 
@@ -221,13 +226,15 @@ export default defineComponent({
     const isPoolCoin = computed(() => {
       return denom.value.startsWith('pool');
     });
+    const stakingEnabled = featureRunning('STAKING');
     const apistore = useStore() as TypedAPIStore;
     const route = useRoute();
     const router = useRouter();
     const denom = computed(() => route.params.denom as string);
 
     pageview({ page_title: 'Asset: ' + route.params.denom, page_path: '/asset/' + route.params.denom });
-    const { balances, balancesByDenom, stakingBalancesByChain, nativeBalances } = useAccount();
+    const { balances, balancesByDenom, stakingBalancesByChain, nativeBalances, unbondingDelegationsByChain } =
+      useAccount();
     const { filterPoolsByDenom, getWithdrawBalances } = usePools();
 
     const assetConfig = computed(() => {
@@ -301,6 +308,13 @@ export default defineComponent({
       }
       return 0;
     });
+    const unbondingDelegation = computed(() => {
+      // TODO: This needs fixing for a chain that supports MULTIPLE stakeable assets (if any ever exist)
+      if (assetConfig.value && assetConfig.value.chain_name && assetConfig.value.stakable) {
+        return unbondingDelegationsByChain(assetConfig.value.chain_name);
+      }
+      return [];
+    });
 
     const stakedAmount = computed(() => {
       let staked = stakingBalance.value;
@@ -316,6 +330,20 @@ export default defineComponent({
       return totalStakedAmount;
     });
 
+    const unstakedAmount = computed(() => {
+      let totalUnstakedAmount = 0;
+      if (unbondingDelegation.value.length > 0) {
+        const unstakedAmounts = unbondingDelegation.value
+          .map((y) => y.entries)
+          .flat()
+          .map((z) => z.balance);
+        if (unstakedAmounts.length > 0) {
+          const unstakedAmount = unstakedAmounts.reduce((acc, item) => +parseInt(item) + acc, 0);
+          totalUnstakedAmount = totalUnstakedAmount + unstakedAmount;
+        }
+      }
+      return totalUnstakedAmount;
+    });
     const poolsInvestedWithAsset = computed(() => {
       const poolsCopy = JSON.parse(JSON.stringify(poolsWithAsset.value));
       const balancesCopy = JSON.parse(JSON.stringify(balances.value));
@@ -370,7 +398,7 @@ export default defineComponent({
     });
 
     const totalAmount = computed(() => {
-      return availableAmount.value + stakedAmount.value;
+      return availableAmount.value + stakedAmount.value + unstakedAmount.value;
     });
 
     const isAreaChartFeatureRunning = featureRunning('PRICE_CHART_ON_ASSET_PAGE') ? true : false;
@@ -442,9 +470,11 @@ export default defineComponent({
       poolsInvestedWithAsset,
       availableAmount,
       stakedAmount,
+      unstakedAmount,
       pooledAmount,
       totalAmount,
       isPoolCoin,
+      stakingEnabled,
       dataStream,
       getTokenPrices,
       showPriceChart,