+++ CHANGELOG.md
@@ -4,6 +4,29 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## 0.1.11 - 2022-02-11
+
+### Features
+
+- Price difference value and percentage implemented (under price)
+
+### Fixes
+
+- Minor fixes on base chart component and asset price performance chart
+
+## 0.1.10 - 2022-02-10
+
+### Features
+
+- Dex Aggregation Exchanges Filter
+- Simplex integration for US
+
+### Fixes
+
+- Display of wrong denom name in asset tables
+- Prevent failure with min pool receive amount
+- Pending transaction modal behavior
+
 ## 0.1.9 - 2022-02-07
 
 ### Features